requirements.txt
comfy-cli
gradio==3.50.2
websockets>=10.0
requests>=2.25.0
Pillow>=8.0.0
pydantic>=2.0.0
civitdl
